+++ CHANGELOG.md
@@ -9,6 +9,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Added
 
+- ![SOURCE] Add source from UN International Children’s Emergency Fund [#95](https://github.com/nbbrd/sdmx-dl/issues/95)
+
 ### Changed
 
 - ![API] Improve code coherence by replacing `Dataflow#of(...)` and `DataQuery#of(...)` with builders